refactor: restructure files to make mvc pattern
