+++ src/main/kotlin/racingcar/Application.kt
@@ -1,23 +1,8 @@
 package racingcar
 
-import racingcar.view.InputView
-import racingcar.view.readRoundCount
-
-
-class Application {
-    companion object {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            val car = InputView.readCarName()
-            val rounds = readRoundCount()
-
-            val game = RacingGame(car)
-            game.play(rounds)
-        }
-    }
-}
+import racingcar.controller.RacingGame
 
 fun main() {
-    Application.main(emptyArray())
+    val game = RacingGame()
+    game.play()
 }
-
