+++ src/main/kotlin/racingcar/model/Car.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.model
 
 import camp.nextstep.edu.missionutils.Randoms
