+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -2,12 +2,10 @@ package racingcar.view
 
 import camp.nextstep.edu.missionutils.Console
 
-object InputView {
-    fun readCarName(): List<String> {
-        print("Enter the names of the cars (comma-separated):\n")
-        val players = Console.readLine() ?: ""
-        return enrollCars(players)
-    }
+fun readCarName(): List<String> {
+    print("Enter the names of the cars (comma-separated):\n")
+    val players = Console.readLine() ?: ""
+    return enrollCars(players)
 }
 
 fun readRoundCount(): Int {
